+++ src/ska_mid_cbf_mcs/controller/controller_device.py
@@ -17,9 +17,9 @@ Sub-element controller device for Mid.CBf
 from __future__ import annotations  # Allows forward references in type hints
 
 import tango
+from ska_control_model import SimulationMode
 from ska_tango_base.base.base_device import DevVarLongStringArrayType
 from ska_tango_base.commands import ResultCode, SubmittedSlowCommand
-from ska_tango_base.control_model import SimulationMode
 from ska_tango_base.utils import convert_dict_to_list
 from tango.server import attribute, command, device_property