+++ src/ska_mid_cbf_mcs/power_switch/pdu_common.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from ska_tango_base.control_model import PowerState
+from ska_control_model import PowerState
 
 
 class Outlet: