+++ src/ska_mid_cbf_mcs/slim/slim_component_manager.py
@@ -19,15 +19,15 @@ from typing import Callable, Optional
 import tango
 import yaml
 from beautifultable import BeautifulTable
-from ska_control_model import TaskStatus
-from ska_tango_base.base.base_component_manager import check_communicating
-from ska_tango_base.commands import ResultCode
-from ska_tango_base.control_model import (
+from ska_control_model import (
     AdminMode,
     HealthState,
     PowerState,
     SimulationMode,
+    TaskStatus,
 )
+from ska_tango_base.base.base_component_manager import check_communicating
+from ska_tango_base.commands import ResultCode
 from ska_tango_testing import context
 
 from ska_mid_cbf_mcs.commons.global_enum import const