+++ src/ska_mid_cbf_mcs/talon_lru/talon_lru_device.py
@@ -17,10 +17,10 @@ from threading import Lock
 from typing import Any
 
 import tango
+from ska_control_model import PowerState, SimulationMode
 from ska_tango_base import SKABaseDevice
 from ska_tango_base.base.base_device import DevVarLongStringArrayType
 from ska_tango_base.commands import ResultCode
-from ska_tango_base.control_model import PowerState, SimulationMode
 from tango.server import attribute, command, device_property
 
 # tango imports